==> Case_study/5-results_merged_solutions.ipynb <==
# Results with Merged Dataset
#### Q5: For all of the models that were produced in 2008 that are still being produced now, how much has the mpg improved and which vehicle improved the most?
Remember to use your new dataset, `combined_dataset.csv`

# load dataset
import pandas as pd
df = pd.read_csv('combined_dataset.csv')

### 1. Create a new dataframe, `model_mpg`, that contain the mean combined mpg values in 2008 and 2018 for each unique model

To do this, group by `model` and find the mean `cmb_mpg_2008` and mean `cmb_mpg` for each.

model_mpg = df.groupby('model').mean()[['cmb_mpg_2008', 'cmb_mpg']]

model_mpg.head()

### 2. Create a new column, `mpg_change`, with the change in mpg
Subtract the mean mpg in 2008 from that in 2018 to get the change in mpg

model_mpg['mpg_change'] = model_mpg['cmb_mpg'] - model_mpg['cmb_mpg_2008']

model_mpg.head()

### 3. Find the vehicle that improved the most
Find the max mpg change, and then use query or indexing to see what model it is!

max_change = model_mpg['mpg_change'].max()
max_change

model_mpg[model_mpg['mpg_change'] == max_change]

Pandas also has a useful [`idxmax`](https://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.idxmax.html) function you can use to find the index of the row containing a column's maximum value!

idx = model_mpg.mpg_change.idxmax()
idx

model_mpg.loc[idx]
